requirements.txt
Flask==3.0.0
Pillow==10.1.0
segno==1.6.1
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0